requirements.txt
fastapi
mangum
uvicorn
requests
pydantic==1.8.2
python-dotenv
openai